Delete old OBM support.
At this point the only externally visible changes are that the obm
extensions are gone, and node_register doesn't accept the obm field
anymore (it was previously silently unused). Internally, this deletes a
whole pile of now-unused code related to our native OBM support; all of
this is handle by OBMd.

This also includes a database migration script which removes unused
columns and tables.

Still TODO, some of the documentation needs updating.